Move OpenTracing shim under ext dir
