+++ ext/opentelemetry-ext-opentracing-shim/src/opentelemetry/ext/opentracingshim/util.py
