+++ tox.ini
@@ -27,7 +27,7 @@ changedir =
   test-ext-jaeger: ext/opentelemetry-ext-jaeger/tests
   test-ext-wsgi: ext/opentelemetry-ext-wsgi/tests
   test-example-app: examples/opentelemetry-example-app/tests
-  test-ot-shim: opentracing-shim/tests
+  test-ot-shim: ext/opentelemetry-ext-opentracing-shim/tests
 
 commands_pre =
 ; Install without -e to test the actual installation
@@ -43,7 +43,7 @@ commands_pre =
   http-requests: pip install {toxinidir}/ext/opentelemetry-ext-http-requests
   jaeger: pip install {toxinidir}/opentelemetry-sdk
   jaeger: pip install {toxinidir}/ext/opentelemetry-ext-jaeger
-  ot-shim: pip install {toxinidir}/opentelemetry-sdk {toxinidir}/opentracing-shim
+  ot-shim: pip install {toxinidir}/opentelemetry-sdk {toxinidir}/ext/opentelemetry-ext-opentracing-shim
 
 ; Using file:// here because otherwise tox invokes just "pip install
 ; opentelemetry-api", leading to an error
@@ -78,7 +78,7 @@ commands_pre =
   pip install -e {toxinidir}/ext/opentelemetry-ext-jaeger
   pip install -e {toxinidir}/ext/opentelemetry-ext-wsgi
   pip install -e {toxinidir}/examples/opentelemetry-example-app
-  pip install -e {toxinidir}/opentracing-shim
+  pip install -e {toxinidir}/ext/opentelemetry-ext-opentracing-shim
 
 commands =
 ; Prefer putting everything in one pylint command to profit from duplication
@@ -95,11 +95,11 @@ commands =
          ext/opentelemetry-ext-http-requests/tests/ \
          ext/opentelemetry-ext-jaeger/src/opentelemetry \
          ext/opentelemetry-ext-jaeger/tests/ \
+         ext/opentelemetry-ext-opentracing-shim/src/ \
+         ext/opentelemetry-ext-opentracing-shim/tests/ \
          ext/opentelemetry-ext-wsgi/tests/ \
          examples/opentelemetry-example-app/src/opentelemetry_example_app/ \
-         examples/opentelemetry-example-app/tests/ \
-         opentracing-shim/src/ \
-         opentracing-shim/tests/
+         examples/opentelemetry-example-app/tests/
   flake8 .
   isort --check-only --diff --recursive .
